requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
configparser==7.2.0
dataclasses-json==0.6.7
distro==1.9.0
docker==7.1.0
frozenlist==1.8.0
google-auth==2.47.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.1.10
langchain-community==0.0.38
langchain-core==0.1.53
langchain-openai==0.0.2
langchain-text-splitters==0.0.2
langsmith==0.1.147
marshmallow==3.26.1
multidict==6.7.0
mypy_extensions==1.1.0
numpy==1.26.4
openai==1.109.1
orjson==3.11.4
packaging==23.2
pandas==2.3.3
pathlib==1.0.1
propcache==0.4.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pywin32==311
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
tenacity==8.5.0
tiktoken==0.5.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
yarl==1.22.0
